requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.1
scikit-learn==1.5.2
joblib==1.3.2
numpy==1.26.4
pandas==2.2.1
pyyaml==6.0.1
requests==2.31.0